modules/sync/requirements.txt
cssselect==0.9.2
lxml==3.6.1
Markdown==2.6.6
Pillow==6.2.0
pymongo==3.3.0
PyPDF2==1.26.0
python-frontmatter==0.3.1
PyYAML==5.1
requests==2.20.0
six==1.10.0
smartcrop==0.1
toml==0.9.2
Unidecode==0.4.19
